requirements.txt
opencv-python>=4.5.5
numpy>=1.21
matplotlib>=3.4
tqdm>=4.60

# For Google Sheets access
gspread>=5.7.0
google-auth>=2.16.0

# For loading .env
python-dotenv>=1.0.0

# For in-memory data handling (optional, if you use Pandas)
pandas>=1.5.0

